requirements.txt
absl-py==0.6.1
astor==0.7.1
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
bleach==1.5.0
blessings==1.7
cycler==0.10.0
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
gast==0.2.0
gpustat==0.5.0
grpcio==1.16.0
h5py==2.8.0
hickle==3.2.2
html5lib==0.9999999
joblib==0.13.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==2.2.3
mock==2.0.0
numpy==1.15.4
nvidia-ml-py==375.53.1
pandas==0.23.4
pbr==5.1.1
pkg-resources==0.0.0
protobuf==3.6.1
psutil==5.4.8
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
root-numpy==4.7.3
root-pandas==0.6.1
scikit-learn==0.20.0
scipy==1.1.0
setGPU==0.0.7
six==1.11.0
sklearn==0.0
subprocess32==3.5.3
tensorboard==1.8.0
tensorflow-gpu==1.8.0
termcolor==1.1.0
Werkzeug==0.14.1
